Fixed compilation issues not resulting in feedback
This was caused by a line of code in RunGame.java that must always
execute, but previously did not execute if there was an exception while
executing the code.

In the future, we would like to give a distinct message if there is a
compilation error rather than just a generic failure.

I also added a script to automatically compile the .java files to .class
and put them in the right location